\name{gcd}
\alias{gcd}

\title{
Great Circle Distance
}
\description{
Takes in 2 pairs of latitute and longitude and gives the distance between them in kms.
}
\usage{
gcd(lat1, long1, lat2, long2)
}

\arguments{
  \item{lat1}{
Latitute of 1st pair
}
  \item{long1}{
Longitude of 1st pair
}
  \item{lat2}{
Latitude of 2nd pair
}
  \item{long2}{
Longitude of 2nd pair
}
}


\value{
Provides the distance between 2 locations in kms
}
\references{
https://en.wikipedia.org/wiki/Great-circle_distance
}
\author{
Anshit Malik
}

